content/canvas/test/test_size.attributes.reflect.1.html

<!DOCTYPE HTML>
<title>Canvas test: size.attributes.reflect.1</title>
<!-- Testing: Setting DOM attributes updates DOM and content attributes -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

canvas.width = 120;
canvas.height = 60;
ok(canvas.getAttribute('width') == '120', "canvas.getAttribute('width') == '120'");
ok(canvas.getAttribute('height') == '60', "canvas.getAttribute('height') == '60'");
ok(canvas.width == 120, "canvas.width == 120");
ok(canvas.height == 60, "canvas.height == 60");

SimpleTest.finish();

});
</script>
